requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
gradio==4.14.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
scipy==1.11.3
matplotlib==3.8.2
